requirements.txt
asgiref==3.11.0
dj-database-url==3.0.1
Django==5.2.8
django-cors-headers==4.9.0
django-filter==25.1
django-otp==1.6.3
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
numpy==2.3.5
pillow==12.0.0
psycopg2-binary==2.9.11
PyJWT==2.10.1
python-decouple==3.8
reportlab==4.2.5
sqlparse==0.5.3
